feat: fix stop car
